Z powodu problemow spowodowanych duza iloscia zmian musze przywrocic aplikacje do stanu z 22.05.12

+++ source/include/server_create_session.h
@@ -13,7 +13,7 @@ Autor: Marcin Kelar ( marcin.kelar@holicon.pl )
 
 void			SESSION_prepare( HTTP_SESSION *http_session );
 void			SESSION_release( HTTP_SESSION *http_session );
-int             SESSION_send_response( HTTP_SESSION *http_session, const char *content_data, int http_content_size );
+short			SESSION_send_response( HTTP_SESSION *http_session, const char *content_data, int http_content_size );
 SEND_INFO*		SESSION_find_response_struct_by_id( int socket );
 HTTP_SESSION*	SESSION_find_by_id( int socket );
 void			SESSION_add_new_send_struct( int socket_descriptor );
